Context on MovieCard without any test
